Fix problems
