+++ server/server.go
@@ -1,21 +1,121 @@
 package main
 
 import (
+	"context"
+	"flag"
+	"fmt"
+	"log"
 	"net/http"
+	"strconv"
+	"time"
 
+	"github.com/go-park-mail-ru/2019_2_TODO/tree/devRK/auth/session"
 	userhttp "github.com/go-park-mail-ru/2019_2_TODO/tree/devRK/server/user/delivery/http"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/naming"
 
 	"github.com/go-park-mail-ru/2019_2_TODO/tree/devRK/server/user/repository"
 	"github.com/go-park-mail-ru/2019_2_TODO/tree/devRK/server/user/usecase"
 	"github.com/go-park-mail-ru/2019_2_TODO/tree/devRK/server/user/utils"
 
+	consulapi "github.com/hashicorp/consul/api"
 	"github.com/labstack/echo"
 	"github.com/labstack/echo/middleware"
 
 	_ "github.com/go-sql-driver/mysql"
 )
 
+var (
+	consulAddr = flag.String("addr", "127.0.0.1:8500", "consul addr (8500 in original consul)")
+)
+
+var (
+	consul       *consulapi.Client
+	nameResolver *testNameResolver
+	servers      []string
+)
+
 func main() {
+	flag.Parse()
+
+	var err error
+	config := consulapi.DefaultConfig()
+	config.Address = *consulAddr
+	consul, err = consulapi.NewClient(config)
+
+	health, _, err := consul.Health().Service("session-api", "", false, nil)
+	if err != nil {
+		log.Fatalf("cant get alive services")
+	}
+
+	servers = []string{}
+	for _, item := range health {
+		addr := item.Service.Address +
+			":" + strconv.Itoa(item.Service.Port)
+		servers = append(servers, addr)
+	}
+
+	nameResolver = &testNameResolver{
+		addr: servers[0],
+	}
+	log.Println(nameResolver)
+
+	grcpConn, err := grpc.Dial(
+		servers[0],
+		grpc.WithInsecure(),
+		grpc.WithBlock(),
+		grpc.WithBalancer(grpc.RoundRobin(nameResolver)),
+	)
+	if err != nil {
+		log.Fatalf("cant connect to grpc")
+	}
+	defer grcpConn.Close()
+
+	if len(servers) > 1 {
+		var updates []*naming.Update
+		for i := 1; i < len(servers); i++ {
+			updates = append(updates, &naming.Update{
+				Op:   naming.Add,
+				Addr: servers[i],
+			})
+		}
+		nameResolver.w.inject(updates)
+	}
+
+	utils.SessManager = session.NewAuthCheckerClient(grcpConn)
+	log.Println(utils.SessManager)
+
+	sess, err := utils.SessManager.Create(context.Background(),
+		&session.Session{
+			Username: "login",
+			Avatar:   "hello",
+		})
+
+	log.Println(sess)
+	if err != nil {
+		log.Println(err)
+	}
+
+	go runOnlineServiceDiscovery(servers)
+
+	go func() {
+		ctx := context.Background()
+		step := 1
+		for {
+			// проверяем несуществуюущую сессию
+			// потому что сейчас между сервисами нет общения
+			// получаем загшулку
+			sess, err := utils.SessManager.Check(ctx,
+				&session.SessionID{
+					ID: "not_exist_" + strconv.Itoa(step),
+				})
+			fmt.Println("get sess", step, sess, err)
+
+			time.Sleep(1500 * time.Millisecond)
+			step++
+		}
+	}()
+
 	e := echo.New()
 
 	// e.Use(middlewares.MetricsMiddleware())
@@ -35,4 +135,52 @@ func main() {
 	userhttp.NewUserHandler(e, usecase.NewUserUsecase(repository.NewUserMemoryRepository()))
 
 	e.Logger.Fatal(e.Start(utils.ListenAddr))
+}
+
+func runOnlineServiceDiscovery(servers []string) {
+	currAddrs := make(map[string]struct{}, len(servers))
+	for _, addr := range servers {
+		currAddrs[addr] = struct{}{}
+	}
+	ticker := time.Tick(5 * time.Second)
+	for _ = range ticker {
+		health, _, err := consul.Health().Service("session-api", "", false, nil)
+		if err != nil {
+			log.Fatalf("cant get alive services")
+		}
+
+		newAddrs := make(map[string]struct{}, len(health))
+		for _, item := range health {
+			addr := item.Service.Address +
+				":" + strconv.Itoa(item.Service.Port)
+			newAddrs[addr] = struct{}{}
+		}
+
+		var updates []*naming.Update
+		// проверяем что удалилось
+		for addr := range currAddrs {
+			if _, exist := newAddrs[addr]; !exist {
+				updates = append(updates, &naming.Update{
+					Op:   naming.Delete,
+					Addr: addr,
+				})
+				delete(currAddrs, addr)
+				fmt.Println("remove", addr)
+			}
+		}
+		// проверяем что добавилось
+		for addr := range newAddrs {
+			if _, exist := currAddrs[addr]; !exist {
+				updates = append(updates, &naming.Update{
+					Op:   naming.Add,
+					Addr: addr,
+				})
+				currAddrs[addr] = struct{}{}
+				fmt.Println("add", addr)
+			}
+		}
+		if len(updates) > 0 {
+			nameResolver.w.inject(updates)
+		}
+	}
 }
