+++ server/user/delivery/http/handlers.go
@@ -3,14 +3,10 @@ package http
 import (
 	"context"
 	"encoding/base64"
-	"flag"
-	"fmt"
 	"io"
 	"log"
 	"net/http"
 	"os"
-	"strconv"
-	"time"
 
 	"github.com/go-park-mail-ru/2019_2_TODO/tree/devRK/auth/session"
 	"github.com/go-park-mail-ru/2019_2_TODO/tree/devRK/server/middlewares"
@@ -18,22 +14,9 @@ import (
 	"github.com/go-park-mail-ru/2019_2_TODO/tree/devRK/server/user"
 	"github.com/go-park-mail-ru/2019_2_TODO/tree/devRK/server/user/utils"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/naming"
 
 	"github.com/labstack/echo"
 	"github.com/microcosm-cc/bluemonday"
-
-	consulapi "github.com/hashicorp/consul/api"
-)
-
-var (
-	consulAddr = flag.String("addr", "127.0.0.1:8500", "consul addr (8500 in original consul)")
-)
-
-var (
-	consul       *consulapi.Client
-	nameResolver *testNameResolver
 )
 
 // Handlers - use UserCRUD
@@ -43,84 +26,6 @@ type Handlers struct {
 
 // NewUserHandler - deliver our handlers in http
 func NewUserHandler(e *echo.Echo, us user.Usecase) {
-	flag.Parse()
-
-	var err error
-	config := consulapi.DefaultConfig()
-	config.Address = *consulAddr
-	consul, err = consulapi.NewClient(config)
-
-	health, _, err := consul.Health().Service("session-api", "", false, nil)
-	if err != nil {
-		log.Fatalf("cant get alive services")
-	}
-
-	servers := []string{}
-	for _, item := range health {
-		addr := item.Service.Address +
-			":" + strconv.Itoa(item.Service.Port)
-		servers = append(servers, addr)
-	}
-
-	nameResolver = &testNameResolver{
-		addr: servers[0],
-	}
-	log.Println(nameResolver)
-
-	grcpConn, err := grpc.Dial(
-		servers[0],
-		grpc.WithInsecure(),
-		grpc.WithBlock(),
-		grpc.WithBalancer(grpc.RoundRobin(nameResolver)),
-	)
-	if err != nil {
-		log.Fatalf("cant connect to grpc")
-	}
-	defer grcpConn.Close()
-
-	if len(servers) > 1 {
-		var updates []*naming.Update
-		for i := 1; i < len(servers); i++ {
-			updates = append(updates, &naming.Update{
-				Op:   naming.Add,
-				Addr: servers[i],
-			})
-		}
-		nameResolver.w.inject(updates)
-	}
-
-	utils.SessManager = session.NewAuthCheckerClient(grcpConn)
-	log.Println(utils.SessManager)
-
-	// go runOnlineServiceDiscovery(servers)
-
-	// ctx := context.Background()
-	// step := 1
-	// for {
-	// 	// проверяем несуществуюущую сессию
-	// 	// потому что сейчас между сервисами нет общения
-	// 	// получаем загшулку
-	// 	sess, err := utils.SessManager.Check(ctx,
-	// 		&session.SessionID{
-	// 			ID: "not_exist_" + strconv.Itoa(step),
-	// 		})
-	// 	fmt.Println("get sess", step, sess, err)
-
-	// 	time.Sleep(1500 * time.Millisecond)
-	// 	step++
-	// }
-
-	sess, err := utils.SessManager.Create(context.Background(),
-		&session.Session{
-			Username: "login",
-			Avatar:   "hello",
-		})
-
-	log.Println(sess)
-	if err != nil {
-		log.Println(err)
-	}
-
 	handlers := Handlers{Users: us}
 
 	e.GET("/metrics", echo.WrapHandler(promhttp.Handler()))
@@ -143,6 +48,7 @@ func (h *Handlers) handlePrometheus(ctx echo.Context) error {
 }
 
 func (h *Handlers) handleSignUp(ctx echo.Context) error {
+	log.Println(utils.SessManager)
 	newUserInput := new(model.User)
 
 	if err := ctx.Bind(newUserInput); err != nil {
@@ -163,9 +69,19 @@ func (h *Handlers) handleSignUp(ctx echo.Context) error {
 	log.Println("Last id: ", lastID)
 	newUserInput.ID = lastID
 
-	if err = utils.SetSession(ctx, *newUserInput); err != nil {
-		ctx.JSON(http.StatusInternalServerError, "Cookie set error")
+	sess, err := utils.SessManager.Create(context.Background(),
+		&session.Session{
+			Username: newUserInput.Username,
+			Avatar:   newUserInput.Avatar,
+		})
+	log.Println(sess)
+	if err != nil {
+		log.Println(err)
 	}
+
+	// if err = utils.SetSession(ctx, newUserInput); err != nil {
+	// 	ctx.JSON(http.StatusInternalServerError, "Cookie set error")
+	// }
 
 	// err = utils.SetToken(ctx)
 	// if err != nil {
@@ -201,7 +117,7 @@ func (h *Handlers) handleSignIn(ctx echo.Context) error {
 	log.Println("UserData: ID - ", userRecord.ID, " Login - ", userRecord.Username,
 		" Avatar - ", userRecord.Avatar)
 
-	if err = utils.SetSession(ctx, *userRecord); err != nil {
+	if err = utils.SetSession(ctx, userRecord); err != nil {
 		ctx.JSON(http.StatusInternalServerError, "Cookie set error")
 	}
 
@@ -297,7 +213,7 @@ func (h *Handlers) handleChangeProfile(ctx echo.Context) error {
 	}
 	log.Println("Update affectedRows: ", affected)
 
-	if err = utils.SetSession(ctx, *changeProfileCredentials); err != nil {
+	if err = utils.SetSession(ctx, changeProfileCredentials); err != nil {
 		ctx.JSON(http.StatusInternalServerError, "Cookie set error")
 	}
 
@@ -341,7 +257,7 @@ func (h *Handlers) handleChangeImage(ctx echo.Context) error {
 	}
 	log.Println("Update affectedRows: ", affected)
 
-	if err = utils.SetSession(ctx, *changeData); err != nil {
+	if err = utils.SetSession(ctx, changeData); err != nil {
 		ctx.JSON(http.StatusInternalServerError, "Cookie set error")
 	}
 
@@ -403,52 +319,4 @@ func (h *Handlers) handleGetProfile(ctx echo.Context) error {
 func (h *Handlers) handleLogout(ctx echo.Context) error {
 	utils.ClearSession(ctx)
 	return ctx.JSON(http.StatusOK, "")
-}
-
-func runOnlineServiceDiscovery(servers []string) {
-	currAddrs := make(map[string]struct{}, len(servers))
-	for _, addr := range servers {
-		currAddrs[addr] = struct{}{}
-	}
-	ticker := time.Tick(5 * time.Second)
-	for _ = range ticker {
-		health, _, err := consul.Health().Service("session-api", "", false, nil)
-		if err != nil {
-			log.Fatalf("cant get alive services")
-		}
-
-		newAddrs := make(map[string]struct{}, len(health))
-		for _, item := range health {
-			addr := item.Service.Address +
-				":" + strconv.Itoa(item.Service.Port)
-			newAddrs[addr] = struct{}{}
-		}
-
-		var updates []*naming.Update
-		// проверяем что удалилось
-		for addr := range currAddrs {
-			if _, exist := newAddrs[addr]; !exist {
-				updates = append(updates, &naming.Update{
-					Op:   naming.Delete,
-					Addr: addr,
-				})
-				delete(currAddrs, addr)
-				fmt.Println("remove", addr)
-			}
-		}
-		// проверяем что добавилось
-		for addr := range newAddrs {
-			if _, exist := currAddrs[addr]; !exist {
-				updates = append(updates, &naming.Update{
-					Op:   naming.Add,
-					Addr: addr,
-				})
-				currAddrs[addr] = struct{}{}
-				fmt.Println("add", addr)
-			}
-		}
-		if len(updates) > 0 {
-			nameResolver.w.inject(updates)
-		}
-	}
 }
